requirements.txt
#best practices to buld in one version and upgrade it later with the newer versions

Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf_spectacular>=0.15.1,<0.16
django-cors-headers==3.13.0